fastapi-app/app/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
pytest
pytest-asyncio
httpx